CLI: Replace `yargs` with `commander` (#10)
* refactor: replace `yargs` wotith `commander`

Also includes some extra features, such as spinners and colored output where supported.

* chore: bump package version

* refactor: create `ttlOnlySpinner` helper

Remove conditionals from the main flow for cleaner code

* refactor(fetchEvents): change parameter order to match other functions in the file

* docs: add JSDoc

* refactor: colorize the entire line depending on status for TABLE format

* refactor: make `--rpc-url` mandatory

* refactor: make `--rpc-url` optional again and default to a public provider

The issue with the default provider from `ethers` is that it tries to obtain consensus from different providers, which usually fail for the kind of query we need to do.

* fix: format README

* chore: fix default options